obstaravania/requirements.txt
argparse
enum34
gensim
htmlmin
jinja2
numpy
paste
psycopg2
python-dateutil
PyYAML
sqlalchemy
webapp2>=3.0.0b1
webob
